# docs/swagger.yaml
basePath: /api
definitions:
...
  server_internal_controller_http_V1_response.NotificationDetail:
    properties:
      comment_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      meta:
        items:
          type: integer
        type: array
      post_slug:
        type: string
      target_url:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  server_internal_controller_http_V1_response.NotificationDetailPage:
    properties:
      current_page:
        type: integer
      list:
        items:
          $ref: '#/definitions/server_internal_controller_http_V1_response.NotificationDetail'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  server_internal_controller_http_V1_response.UnreadCount:
    properties:
      count:
        type: integer
    type: object
...
  server_internal_controller_http_shared.Envelope:
    properties:
      code:
        example: "0000"
        type: string
      data: {}
      message:
        example: ok
        type: string
    type: object
info:
  contact: {}
  description: Nlxiaoy Blog 后端 API（Admin + Public V1）。
  title: Nlxiaoy Blog API
  version: "1.0"
paths:
...
  /v1/notifications:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 分页大小
        in: query
        name: page_size
        type: integer
      - description: 排序字段
        enum:
        - created_at
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 排序方向
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: 是否已读过滤
        in: query
        name: filter.is_read
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/server_internal_controller_http_V1_response.NotificationDetailPage'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 通知列表
      tags:
      - V1.Notifications
  /v1/notifications/{id}:
    delete:
      parameters:
      - description: 通知 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 删除通知
      tags:
      - V1.Notifications
  /v1/notifications/{id}/read:
    put:
      parameters:
      - description: 通知 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 标记通知已读
      tags:
      - V1.Notifications
  /v1/notifications/read-all:
    put:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 标记全部通知已读
      tags:
      - V1.Notifications
  /v1/notifications/stream:
    get:
      description: 使用 Query Token 鉴权（token=access_token），返回 text/event-stream。
      parameters:
      - description: Access Token
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      summary: 通知 SSE 流
      tags:
      - V1.Notifications
  /v1/notifications/unread:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/server_internal_controller_http_V1_response.UnreadCount'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 未读通知数量
      tags:
      - V1.Notifications
...
schemes:
- http
- https
securityDefinitions:
  AdminSession:
    in: cookie
    name: fiber_session
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"